Inline reduce.c
This commits inlines reduce.c entirely. The goal is to get performance without
-flto closer to performance with -flto.
Namespacing will be added in the next commit.

Note that some of these functions are a potential side-channel leak.
That needs to be addressed urgently, but it with -flto it was already present
prior to this commit.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>