CBMC: Remove CBMC_OBJECT_BITS from all proof Makefiles and default to 12
I'm tired of constantly debugging CBMC proofs just to find out that some
unrelated proof ran out of addressing space and needs the OBJECT_BITS to
be bumped from 8 to 10.
This commit changes the default to 12, which is the largest value that we
are currently using, and removes the CBMC_OBJECT_BITS from the individual
Makefiles.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>