HashML-DSA: Replace enum by define to pick pre-hash function
Currently we have an enum for the 12 different pre-hash functions for
HashML-DSA. This leads to problems in multi-level builds as we must only
define the enum once. Currently we work around this by guarding the enum
definition with a pre-processor conditional.
However, there is also a (theoretical) concern about the type of the enum
being implementation-defined in C90:
https://github.com/pq-code-package/mldsa-native/pull/537#discussion_r2484226645

It seems cleaner to not use an enum here, but instead use #defines avoiding all
the above problems.
This commit implements that change.

We also eliminate the camel case hashAlg - that was inconsistent with the
remaining code base from the start.

Resolves
https://github.com/pq-code-package/mldsa-native/issues/591

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>