test: Add --exclude-example flag and use to skip basic_determinitic with PCT
MLD_CONFIG_NO_RANDOMIZED_API and MLD_CONFIG_KEYGEN_PCT are incompatible and
hence the config variation test that enables PCT currently errors in the
basic_deterministic example (with the appopriate error message that the options
are incompatible).

This commit ports the --exclude-example option for the test script and uses
it in CI for the PCT test.

We also consolidate various small differences to the mlkem-native test script
(whitespace changes, etc.).

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>