nix: Only run build_nix_cache CI in main branch
b4209badf3bb948a626dbfde8e5e789f947323d2 changed the nix CI to only store the
cache in the main branch to stop PRs from evicting our cache entries from
main.
However, with that commit PRs changing the nix setup are still running the
build_nix_cache CI, but discarding the result in the end.
That is not useful and wastes significant CI time.
This commit changes this to only run the build_nix_cache workflow in the main
branch.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>