Add runtime dispatch (mld_poly_chknorm)
The inputs to chknorm are potentially secret. However, it is fine (and
unavoidable) to leak if chknorm return 0 or 1 (i.e., if all coeffs are within
bound or not). The declassification of that currently happens in sign.c.
However, for the run-time dispatch we require to branch depending on whether
mld_poly_chknorm_native returns MLD_NATIVE_FUNC_FALLBACK (-1) or not to signal
that the platform does not have the required capabilities to run the native
code (and we should hence fallback to the C code).

This commit adds a declassification of (ret == FALLBACK) which is sufficient
to make this code pass the constant time tests. The declassification of the
actual value (0/1) remains in sign.c to be consistent with the C
implementation.

Co-authored-by: Matthias J. Kannwischer <matthias@kannwischer.eu>
Signed-off-by: willieyz <willie.zhao@chelpis.com>