x86_64: Avoid overread in polyz_unpack
polyz_unpack_17_avx2/polyz_unpack_197_avx2 unpacks polynnomials with
coefficients packed into 18/20 bits each.
To do so, it currently loads 32 bytes into a 256-bit register and discards
the last 14/12-bytes. This is problematic in the last iteration as this
overreads the buffer presenting a potential safety problem.

This is a violoation of the API contract which only requires 18*32/20*32-bytes.
Hence, this commit eliminates the overread by adding a special handling for
the last iteration.

Luckily, this does not presenta problem in the current code base as all call
sites have sufficient padding (by chance):
1) In mld_poly_uniform_gamma1, the buffer is padded to full SHAKE256 blocks,
   i.e., for MLDSA_POLYZ_PACKEDBYTES=576 => ceil(576/136)*136=680, and
             MLDSA_POLYZ_PACKEDBYTES=640 => ceil(640/136)*136=680.
2) In mld_unpack_sig where the signature is unpacked, the z component is
   followed by the hint component which is 83 bytes.

This means both overreads are fine.

We did, however, find this problem in the unit tests that are being implemented
in https://github.com/pq-code-package/mldsa-native/pull/777.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>
Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>